+++ src/frequenz/sdk/actor/_power_managing/_power_managing_actor.py
@@ -11,7 +11,7 @@ import typing
 from datetime import datetime, timedelta, timezone
 
 from frequenz.channels import Receiver, Sender
-from frequenz.channels.util import select, selected_from
+from frequenz.channels.util import SkipMissedAndDrift, Timer, select, selected_from
 from typing_extensions import override
 
 from ...timeseries._base_types import PoolType, SystemBounds
@@ -75,7 +75,9 @@ class PowerManagingActor(Actor):
         self._subscriptions: dict[frozenset[int], dict[int, Sender[_Report]]] = {}
         self._distribution_results: dict[frozenset[int], power_distributing.Result] = {}
 
-        self._algorithm: BaseAlgorithm = Matryoshka()
+        self._algorithm: BaseAlgorithm = Matryoshka(
+            max_proposal_age=timedelta(seconds=60.0)
+        )
 
         super().__init__()
 
@@ -180,10 +182,13 @@ class PowerManagingActor(Actor):
     @override
     async def _run(self) -> None:
         """Run the power managing actor."""
+        last_result_partial_failure = False
+        drop_old_proposals_timer = Timer(timedelta(seconds=1.0), SkipMissedAndDrift())
         async for selected in select(
             self._proposals_receiver,
             self._bounds_subscription_receiver,
             self._power_distributing_results_receiver,
+            drop_old_proposals_timer,
         ):
             if selected_from(selected, self._proposals_receiver):
                 proposal = selected.value
@@ -234,9 +239,20 @@ class PowerManagingActor(Actor):
                 self._distribution_results[
                     frozenset(result.request.component_ids)
                 ] = result
+                if not isinstance(result, power_distributing.Success):
+                    _logger.warning(
+                        "PowerManagingActor: PowerDistributing failed: %s", result
+                    )
                 match result:
                     case power_distributing.PartialFailure(request):
-                        await self._send_updated_target_power(
-                            frozenset(request.component_ids), None, must_send=True
-                        )
+                        if not last_result_partial_failure:
+                            last_result_partial_failure = True
+                            await self._send_updated_target_power(
+                                frozenset(request.component_ids), None, must_send=True
+                            )
+                    case power_distributing.Success():
+                        last_result_partial_failure = False
                 await self._send_reports(frozenset(result.request.component_ids))
+
+            elif selected_from(selected, drop_old_proposals_timer):
+                self._algorithm.drop_old_proposals(asyncio.get_event_loop().time())